add: 기존 lagacy Controller을 LagacyHandlerExecution으로 wrapping함
