+++ src/main/java/core/mvc/asis/RequestMapping.java
@@ -1,6 +1,8 @@
 package core.mvc.asis;
 
+import core.mvc.tobe.HandlerExecution;
 import core.mvc.tobe.HandlerMapping;
+import core.mvc.tobe.LagacyHandlerExecution;
 import next.controller.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -40,7 +42,13 @@ public class RequestMapping implements HandlerMapping {
     }
 
     @Override
-    public Object getHandler(HttpServletRequest request) {
-        return findController(request.getRequestURI());
+    public HandlerExecution getHandler(HttpServletRequest request) {
+        Controller controller = findController(request.getRequestURI());
+
+        if (controller != null) {
+            return new LagacyHandlerExecution(controller);
+        }
+
+        return null;
     }
 }
