+++ src/main/java/core/mvc/tobe/HandlerMapping.java
@@ -6,5 +6,5 @@ import javax.servlet.http.HttpServletRequest;
  * @author KingCjy
  */
 public interface HandlerMapping {
-    Object getHandler(HttpServletRequest request);
+    HandlerExecution getHandler(HttpServletRequest request);
 }
