+++ src/main/java/core/mvc/tobe/LagacyHandlerExecution.java
@@ -1,0 +1,23 @@
+package core.mvc.tobe;
+
+import core.mvc.ModelAndView;
+import core.mvc.asis.Controller;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+public class LagacyHandlerExecution implements HandlerExecution {
+
+    private Controller controller;
+
+    public LagacyHandlerExecution(Controller controller) {
+        this.controller = controller;
+    }
+
+    @Override
+    public ModelAndView handle(HttpServletRequest request, HttpServletResponse response) throws Exception {
+        String viewName = controller.execute(request, response);
+
+        return viewName != null ? new ModelAndView(viewName) : null;
+    }
+}
